+++ packages/orm/angel_orm_postgres/lib/src/orm_postgres_pool.dart
@@ -34,7 +34,9 @@ class PostgreSqlPoolExecutor extends QueryExecutor {
   @override
   Future<Result> query(
       String tableName, String query, Map<String, dynamic> substitutionValues,
-      {String returningQuery = '', List<String> returningFields = const []}) {
+      {String returningQuery = '',
+      String resultQuery = '',
+      List<String> returningFields = const []}) {
     if (returningFields.isNotEmpty) {
       var fields = returningFields.join(', ');
       var returning = 'RETURNING $fields';