+++ packages/orm/angel_orm_postgres/test/belongs_to_test.dart
@@ -11,7 +11,7 @@ import 'util.dart';
 void main() {
   Logger.root.level = Level.ALL; // defaults to Level.INFO
   Logger.root.onRecord.listen((record) {
-    print('${record.level.name}: ${record.time}: ${record.message}');
+    print('${record.loggerName}: ${record.time}: ${record.message}');
   });
 
   late Connection conn;
@@ -43,7 +43,12 @@ void main() {
     deathlyHallows = (await bookQuery.insert(executor)).value;
   });
 
-  tearDown(() async => await dropTables(runner));
+  tearDown(() async {
+    await dropTables(runner);
+    if (conn.isOpen) {
+      await conn.close();
+    }
+  });
 
   group('selects', () {
     test('select all', () async {