+++ packages/orm/angel_orm_postgres/test/enum_and_nested_test.dart
@@ -9,7 +9,7 @@ import 'models/has_car.dart';
 void main() {
   Logger.root.level = Level.ALL; // defaults to Level.INFO
   Logger.root.onRecord.listen((record) {
-    print('${record.level.name}: ${record.time}: ${record.message}');
+    print('${record.loggerName}: ${record.time}: ${record.message}');
   });
 
   late Connection conn;
@@ -24,6 +24,9 @@ void main() {
 
   tearDown(() async {
     await dropTables(runner);
+    if (conn.isOpen) {
+      await conn.close();
+    }
   });
 
   test('insert', () async {