+++ packages/orm/angel_orm_postgres/test/has_map_test.dart
@@ -9,7 +9,7 @@ import 'models/has_map.dart';
 void main() {
   Logger.root.level = Level.ALL; // defaults to Level.INFO
   Logger.root.onRecord.listen((record) {
-    print('${record.level.name}: ${record.time}: ${record.message}');
+    print('${record.loggerName}: ${record.time}: ${record.message}');
   });
 
   late Connection conn;
@@ -22,7 +22,12 @@ void main() {
     runner = await createTables(conn, [HasMapMigration()]);
   });
 
-  tearDown(() async => await dropTables(runner));
+  tearDown(() async {
+    await dropTables(runner);
+    if (conn.isOpen) {
+      await conn.close();
+    }
+  });
 
   test('insert', () async {
     var query = HasMapQuery();