+++ packages/orm/angel_orm_mysql/CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 8.3.1
 
 * Updated test cases
-* Fixed deletion to return deleted records
+* Updated deletion to return deleted record(s)
+* Fixed select and insert query returning incorrect data when the joined tables have same field names
+* Fixed issues due to breaking changes in `analyzer` 7.3.0
 
 ## 8.3.0