+++ packages/serialize/angel_serialize_generator/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 8.4.1
 
-* Fixed `ShimFieldImpl` due to change of `type` from `DartType` to `TypeImpl`
+* Fixed `ShimFieldImpl` due to breaking changes in `analyzer` 7.3.0
 
 ## 8.4.0